refactored module structure, cleaned up imports (QueryStrategies)
