+++ src/QueryStrategies/qs_base.jl
@@ -1,5 +1,20 @@
+abstract type QueryStrategy end
+abstract type PoolQs <: QueryStrategy end
 
-function initialize_qs(qs, model::OCClassifier, data::Array{T, 2}, params)::qs where T <: Real
+function qs_score end
+
+"""
+get_query_object(qs::QueryStrategy, data::Array{T, 2}, pools::Vector{Symbol}, global_indices::Vector{Int}, history::Vector{Int})
+
+# Arguments
+- `query_data`: Subset of the full data set
+- `pools`: Labels for `query_data`
+- `global_indices`: Indices of the observations in `query_data` relative to the full data set.
+- `history`: Indices of previous queries
+"""
+function get_query_object end
+
+function initialize_qs(qs::DataType, model::SVDD.OCClassifier, data::Array{T, 2} where T <: Real, params)::qs
     if qs <: HybridPQs || qs <: HybridQss
         return qs(model, data; params...)
     elseif qs <: ModelBasedPQs || qs <: ModelBasedQss
